requirements.txt
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
tkinterdnd2==0.3.0
psutil==5.9.0
tkvideoplayer==2.3
